Move dynamic preload code to renamed package
