+++ collect_app/src/main/java/org/odk/collect/android/dynamicpreload/ExternalAnswerResolver.java
@@ -16,7 +16,7 @@
  * the License.
  */
 
-package org.odk.collect.android.externaldata;
+package org.odk.collect.android.dynamicpreload;
 
 import org.javarosa.core.model.Constants;
 import org.javarosa.core.model.FormDef;
