+++ collect_app/src/main/java/org/odk/collect/android/dynamicpreload/ExternalAppsUtils.java
@@ -16,7 +16,7 @@
  * the License.
  */
 
-package org.odk.collect.android.externaldata;
+package org.odk.collect.android.dynamicpreload;
 
 import android.content.Intent;
