+++ collect_app/src/main/java/org/odk/collect/android/dynamicpreload/ExternalDataReader.java
@@ -16,7 +16,7 @@
  * the License.
  */
 
-package org.odk.collect.android.externaldata;
+package org.odk.collect.android.dynamicpreload;
 
 import java.io.File;
 import java.util.Map;
