+++ collect_app/src/main/java/org/odk/collect/android/dynamicpreload/ExternalSelectChoice.java
@@ -16,16 +16,28 @@
  * the License.
  */
 
-package org.odk.collect.android.externaldata;
+package org.odk.collect.android.dynamicpreload;
 
-import org.javarosa.core.model.condition.IFunctionHandler;
+import org.javarosa.core.model.SelectChoice;
 
 /**
  * Author: Meletis Margaritis
- * Date: 30/04/13
- * Time: 09:42
+ * Date: 18/11/2013
+ * Time: 2:08 μμ
  */
-public interface ExternalDataHandler extends IFunctionHandler {
+public class ExternalSelectChoice extends SelectChoice {
 
-    void setExternalDataManager(ExternalDataManager externalDataManager);
+    String image;
+
+    public ExternalSelectChoice(String labelOrID, String value, boolean isLocalizable) {
+        super(labelOrID, value, isLocalizable);
+    }
+
+    public String getImage() {
+        return image;
+    }
+
+    public void setImage(String image) {
+        this.image = image;
+    }
 }
