+++ collect_app/src/main/java/org/odk/collect/android/dynamicpreload/handler/ExternalDataHandlerPull.java
@@ -16,7 +16,7 @@
  * the License.
  */
 
-package org.odk.collect.android.externaldata.handler;
+package org.odk.collect.android.dynamicpreload.handler;
 
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
@@ -24,9 +24,9 @@ import android.database.sqlite.SQLiteException;
 
 import org.javarosa.core.model.condition.EvaluationContext;
 import org.javarosa.xpath.expr.XPathFuncExpr;
-import org.odk.collect.android.externaldata.ExternalDataManager;
-import org.odk.collect.android.externaldata.ExternalDataUtil;
-import org.odk.collect.android.externaldata.ExternalSQLiteOpenHelper;
+import org.odk.collect.android.dynamicpreload.ExternalDataManager;
+import org.odk.collect.android.dynamicpreload.ExternalDataUtil;
+import org.odk.collect.android.dynamicpreload.ExternalSQLiteOpenHelper;
 
 import java.util.ArrayList;
 import java.util.List;
