+++ collect_app/src/main/java/org/odk/collect/android/tasks/FormLoaderTask.java
@@ -39,10 +39,10 @@ import org.javarosa.xform.util.XFormUtils;
 import org.javarosa.xpath.XPathTypeMismatchException;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.application.initialization.ContainsSearchOrPullDataExtra;
-import org.odk.collect.android.externaldata.ExternalAnswerResolver;
-import org.odk.collect.android.externaldata.ExternalDataManager;
-import org.odk.collect.android.externaldata.ExternalDataReader;
-import org.odk.collect.android.externaldata.ExternalDataReaderImpl;
+import org.odk.collect.android.dynamicpreload.ExternalAnswerResolver;
+import org.odk.collect.android.dynamicpreload.ExternalDataManager;
+import org.odk.collect.android.dynamicpreload.ExternalDataReader;
+import org.odk.collect.android.dynamicpreload.ExternalDataReaderImpl;
 import org.odk.collect.android.fastexternalitemset.ItemsetDbAdapter;
 import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.javarosawrapper.JavaRosaFormController;
