+++ collect_app/src/test/java/org/odk/collect/android/dynamicpreload/ExternalDataUtilTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.externaldata;
+package org.odk.collect.android.dynamicpreload;
 
 import org.junit.Assert;
 import org.junit.Test;
